requirements.txt
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
beautifulsoup4==4.14.3
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
dj-database-url==3.0.1
Django==4.2.17
django-bootstrap-v5==1.0.11
django-elasticsearch-dsl==9.0
django-filter==24.3
djangorestframework==3.16.1
elastic-transport==9.2.0
elasticsearch==9.2.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
httplib2==0.31.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.33.2
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.1
zope.interface==8.1.1
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0django-storages==1.14.6
boto3==1.42.37
